Optimized addWeaponGroup()
Very often, when this chip is used with multiple turrets and guns, a simple approach of looping through the acf links of all guns is extremely inefficient as in the vast majority of these cases guns in the same weapon group will be linked to the same ammo crates. We can add some logic which saves a value onto the ammo crate which specifies which weapon groups the ammo crate is already a part of, accessing this value later has a time cost of O(1). We can use this value to skip ammo crates that have already been processed by the function and gain massive performance when large inputs are provided to the addWeaponGroup() function. This is extremely significant as this function was the main bottleneck of this chip, causing "tick quota exceeded" crashes upon init.